Revert "Performance optimizations to seeding (using Enumerated Radix Trees) and mate-rescue"

+++ src/kthread.cpp
@@ -62,21 +62,22 @@ static void *ktf_worker(void *data)
 	
 	for (;;) {
 		i = __sync_fetch_and_add(&w->i, w->t->n_threads);
-		long st = i * BATCH_SIZE;
+		int st = i * BATCH_SIZE;
 		if (st >= w->t->n) break;
-		long ed = (i + 1) * BATCH_SIZE < w->t->n? (i + 1) * BATCH_SIZE : w->t->n;
-		w->t->func(w->t->data, st, ed-st, w - w->t->w);
+		int ed = (i + 1) * BATCH_SIZE < w->t->n? (i + 1) * BATCH_SIZE : w->t->n;
+		// w->t->func(w->t->data, st, ed-st, tid);
+        w->t->func(w->t->data, st, ed-st, w - w->t->w);
 	}
 
 	while ((i = steal_work(w->t)) >= 0) {
-		long st = i * BATCH_SIZE;
-		long ed = (i + 1) * BATCH_SIZE < w->t->n? (i + 1) * BATCH_SIZE : w->t->n;
+		int st = i * BATCH_SIZE;
+		int ed = (i + 1) * BATCH_SIZE < w->t->n? (i + 1) * BATCH_SIZE : w->t->n;
 		w->t->func(w->t->data, st, ed-st, w - w->t->w);
 	}
 	pthread_exit(0);
 }
 
-void kt_for(void (*func)(void*, long, long, int), void *data, int n)
+void kt_for(void (*func)(void*, int, int, int), void *data, int n)
 {
 	int i;
 	kt_for_t t;
@@ -91,7 +92,7 @@ void kt_for(void (*func)(void*, long, long, int), void *data, int n)
 		t.w[i].t = &t, t.w[i].i = i;
 
 	pthread_attr_t attr;
-	pthread_attr_init(&attr);
+    pthread_attr_init(&attr);
 	
 	// printf("getcpu: %d\n", sched_getcpu());
 	for (i = 0; i < t.n_threads; ++i) {
